Add logging to jobs
In order to get a consistent ILogger prefix across log statements inside
a jobs job fragments, we need to pass a logger through. This was done by
adding a public logger to the root job and then passing the job through
to the job fragments 'Run' function as context. This allows all job
fragments inside a job to access the same logger.